+++ src/stan/math/matrix/get_base1_lhs.hpp
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <stan/math/matrix/Eigen.hpp>
-#include <stan/math/matrix/check_range.hpp>
+#include <stan/math/error_handling/matrix/check_range.hpp>
 
 namespace stan {
   namespace math {
