+++ src/stan/math/matrix_error_handling.hpp
@@ -16,6 +16,7 @@
 #include <stan/math/error_handling/matrix/check_pos_definite.hpp>
 #include <stan/math/error_handling/matrix/check_pos_semidefinite.hpp>
 #include <stan/math/error_handling/matrix/check_positive_ordered.hpp>
+#include <stan/math/error_handling/matrix/check_range.hpp>
 #include <stan/math/error_handling/matrix/check_row_index.hpp>
 #include <stan/math/error_handling/matrix/check_simplex.hpp>
 #include <stan/math/error_handling/matrix/check_size_match.hpp>
