--- src/test/unit/math/matrix/check_range_test.cpp
@@ -1,2 +1,0 @@
-#include <stan/math/matrix/check_range.hpp>
-#include <gtest/gtest.h>
